Clearbit API wrapper for getting logo for domain
The Clearbit API wrapper has been updated to have a method for accessing
the Clearbit Logo API which takes a domain name and returns the image
data for the logo in PNG format.

Our wrapper then converts this PNG data into a valid data URL by
encoding it as base64.